+++ notifications/WingspanNotifications/BenefitsService.cs
@@ -49,12 +49,12 @@ namespace WingspanNotifications
     /// </summary>
     public class BenefitsService: IBenefitsService
     {
-        public SDKConfig Config { get; private set; }
+        public SDKConfig SDKConfiguration { get; private set; }
         private const string _language = "csharp";
-        private const string _sdkVersion = "2.0.1";
-        private const string _sdkGenVersion = "2.194.1";
+        private const string _sdkVersion = "2.0.2";
+        private const string _sdkGenVersion = "2.209.0";
         private const string _openapiDocVersion = "1.0.0";
-        private const string _userAgent = "speakeasy-sdk/csharp 2.0.1 2.194.1 1.0.0 WingspanNotifications";
+        private const string _userAgent = "speakeasy-sdk/csharp 2.0.2 2.209.0 1.0.0 WingspanNotifications";
         private string _serverUrl = "";
         private ISpeakeasyHttpClient _defaultClient;
         private ISpeakeasyHttpClient _securityClient;
@@ -64,13 +64,13 @@ namespace WingspanNotifications
             _defaultClient = defaultClient;
             _securityClient = securityClient;
             _serverUrl = serverUrl;
-            Config = config;
+            SDKConfiguration = config;
         }
         
 
         public async Task<GetBenefitsServiceResponse> GetBenefitsServiceAsync()
         {
-            string baseUrl = this.Config.GetTemplatedServerDetails();
+            string baseUrl = this.SDKConfiguration.GetTemplatedServerDetails();
             var urlString = baseUrl + "/benefits/service";
             
             var httpRequest = new HttpRequestMessage(HttpMethod.Get, urlString);
@@ -110,7 +110,7 @@ namespace WingspanNotifications
                 Id = id,
                 ServiceEnablementUpdate = serviceEnablementUpdate,
             };
-            string baseUrl = this.Config.GetTemplatedServerDetails();
+            string baseUrl = this.SDKConfiguration.GetTemplatedServerDetails();
             var urlString = URLBuilder.Build(baseUrl, "/benefits/service/{id}", request);
             
             var httpRequest = new HttpRequestMessage(HttpMethod.Patch, urlString);