+++ notifications/RELEASES.md
@@ -30,3 +30,13 @@ Based on:
 - [csharp v2.0.1] notifications
 ### Releases
 - [NuGet v2.0.1] https://www.nuget.org/packages/WingspanNotifications/2.0.1 - notifications
+
+## 2023-12-04 01:28:04
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.125.0 (2.209.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v2.0.2] notifications
+### Releases
+- [NuGet v2.0.2] https://www.nuget.org/packages/WingspanNotifications/2.0.2 - notifications